Move HTTP tracing into new tracing package
